requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
crispy-bootstrap5==2024.10
django-crispy-forms==2.3
python-decouple==3.8
whitenoise==6.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
boto3==1.35.75
django-storages==1.14.4
qrcode==8.0
python-dateutil==2.9.0.post0
dj-database-url==2.2.0

faker==37.5.3

